+++ python-libraries/data-platform-catalogue/data_platform_catalogue/__init__.py
@@ -1,3 +1,8 @@
 from .client import CatalogueError, ReferencedEntityMissing  # noqa: F401
-from .entities import DataProductMetadata  # noqa: F401
-from .entities import CatalogueMetadata, DataLocation, TableMetadata  # noqa: F401
+from .entities import (  # noqa: F401
+    CatalogueMetadata,
+    DatabaseMetadata,
+    DataLocation,
+    DataProductMetadata,
+    TableMetadata,
+)